updated Track constructor

+++ Track.cpp
@@ -1,6 +1,6 @@
 #include "Track.h"
 
-Track::Track(string _name, string _artist, int _popularity, int _xplicit, int _releaseYear, int _danceability, int _energy, int _valence) {
+Track::Track(string _name, string _artist, int _popularity, int _xplicit, int _releaseYear, float _danceability, float _energy, float _valence) {
     name = _name;
     artist = _artist;
     popularity = _popularity;
