+++ Track.h
@@ -10,7 +10,7 @@ private:
     float danceability, energy, valence;
 public:
     //constructor
-    Track(string _name, string _artist, int _popularity, int _xplicit, int _releaseYear, int _danceability, int _energy, int _valence);
+    Track(string _name, string _artist, int _popularity, int _xplicit, int _releaseYear, float _danceability, float _energy, float _valence);
 
     //accessors
     string getName();
